# perf(context): improve performance of context (setHeaders)