@EdamAme-x 2024-10-31
Benchmark Source

```ts
import { run, bench } from 'mitata'

const subHeaders = {
  a: '1',
  b: '2',
  c: '3',
  d: '4',
  e: '5',
  f: '6',
}

bench('using .entries', () => {
  const headers = new Headers()

  Object.entries(subHeaders).forEach(([key, value]) => headers.set(key, value))
})

bench('using for syntax', () => {
  const headers = new Headers()

  for (let i = 0, keys = Object.keys(subHeaders), len = keys.length; i < len; i++) {
    const key = keys[i]
    headers.set(key, subHeaders[key])
  }
})

await run()
```